+++ src/models/Message.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace mutation\translate\models;
+
+use Yii;
+use yii\db\ActiveRecord;
+
+/**
+ * This is the model class for table "message".
+ *
+ * @property integer $id
+ * @property string $language
+ * @property string $translation
+ *
+ * @property SourceMessage $id0
+ */
+class Message extends ActiveRecord
+{
+    /**
+     * @inheritdoc
+     */
+    public static function tableName()
+    {
+        return '{{%message}}';
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function rules()
+    {
+        return [
+            [['id', 'language'], 'required'],
+            [['id'], 'integer'],
+            [['translation'], 'string'],
+            [['language'], 'string', 'max' => 16],
+            [['id'], 'exist', 'skipOnError' => true, 'targetClass' => SourceMessage::className(), 'targetAttribute' => ['id' => 'id']],
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function attributeLabels()
+    {
+        return [
+            'id' => 'ID',
+            'language' => 'Language',
+            'translation' => 'Translation',
+        ];
+    }
+}